protect seed against production environment

+++ db/seeds.rb
@@ -7,23 +7,25 @@
 #   cities = City.create([{ :name => 'Chicago' }, { :name => 'Copenhagen' }])
 #   Mayor.create(:name => 'Daley', :city => cities.first)
 
-User.delete_all
-user = User.create!({
-  :name => "Quentin Tarantino",
-  :email => "quentin@example.com",
-  :password => "test123"
-})
-user.confirm!
-
-Food.delete_all
-lines = File.open(Rails.root.join('db','foods.txt')).readlines
-lines.each do |line|
-  p = line.split("|")
-  Food.create!({
-    :name => p[0],
-    :weight => p[1],
-    :measure => p[2],
-    :kcal => p[3],
-    :kind => p[4],
+if Rails.env.development?
+  User.delete_all
+  user = User.create!({
+    :name => "Quentin Tarantino",
+    :email => "quentin@example.com",
+    :password => "test123"
   })
+  user.confirm!
+
+  Food.delete_all
+  lines = File.open(Rails.root.join('db','foods.txt')).readlines
+  lines.each do |line|
+    p = line.split("|")
+    Food.create!({
+      :name => p[0],
+      :weight => p[1],
+      :measure => p[2],
+      :kcal => p[3],
+      :kind => p[4],
+    })
+  end
 end
